third_party/blink/web_tests/http/tests/security/sandbox-inherit-to-initial-document-2-expected.txt

CONSOLE ERROR: Blocked script execution in 'about:blank' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Test how sandbox flags get inherited to initial document in subframe.

PASS
